requirements.txt
flask
flask_sqlalchemy
g4f
langdetect
